Add files via upload
"The Simon Game is a classic electronic memory game brought to life as a web application. Built using HTML, CSS, and JavaScript, this project offers an interactive and engaging experience for players of all ages. The game challenges players to repeat a series of randomly generated patterns of colors and sounds, testing their memory and concentration skills. With its user-friendly interface and responsive design, the Simon Game provides hours of entertainment and fun. Try your hand at this nostalgic game and see if you can beat the high score!"